requirements.txt
hyperopt==0.2.5
ipykernel==6.2.0
mlflow==1.20.2
pyarrow==5.0.0
scikit-learn==0.24.2
dagshub==0.1.8
hydra-core==1.2.0
numpy==1.22.4
pandas==1.4.2
streamlit==1.9.0
xgboost==1.6.1
joblib==1.1.0